stack v1.02 - Corrections to move assignment operator, trim_trailing_groups, emplace, change_group_sizes and test_suite.

+++ SG14_test/plf_benchmark_suite/plf_bench.h
@@ -1440,7 +1440,7 @@ inline PLF_FORCE_INLINE void benchmark_general_use_percentage(const unsigned int
 			}
 		}
 
-		end_approximate_memory_use += container.approximate_memory_use();
+		end_approximate_memory_use += static_cast<unsigned int>(container.approximate_memory_use());
 	}
 	
 	end_approximate_memory_use /= (number_of_runs / 10) + 1;
@@ -1561,7 +1561,7 @@ inline PLF_FORCE_INLINE void benchmark_general_use_small_percentage(const unsign
 			}
 		}
 
-		end_approximate_memory_use += container.approximate_memory_use();
+		end_approximate_memory_use += static_cast<unsigned int>(container.approximate_memory_use());
 	}
 
 	end_approximate_memory_use /= dump_run_end;
